requirements.txt
asgiref==3.11.1
click==8.3.1
dj-database-url==3.1.2
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
gunicorn==25.1.0
h11==0.16.0
packaging==26.0
psycopg2-binary==2.9.11
sqlparse==0.5.5
uvicorn==0.41.0
